+++ frontend/src/apis/checklist.ts
@@ -1,6 +1,6 @@
 import fetcher from '@/apis/fetcher';
 import { BASE_URL, ENDPOINT } from '@/apis/url';
-import { Answer } from '@/pages/ChecklistPage';
+import { Answer } from '@/pages/NewChecklistPage';
 
 export const getChecklistQuestions = async () => {
   const response = await fetcher.get({ url: BASE_URL + ENDPOINT.CHECKLIST_QUESTION });
