+++ frontend/src/components/NewChecklist/ChecklistQuestion.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 import { QuestionDot } from '@/assets/assets';
 import FaceIcon from '@/components/common/FaceMark/FaceIcon';
-import { addAnswerProps } from '@/pages/ChecklistPage';
+import { addAnswerProps } from '@/pages/NewChecklistPage';
 import { ChecklistQuestion } from '@/types/checklist';
 
 interface Props {
