+++ frontend/src/components/NewChecklist/OptionModal/OptionModalInfoBox.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
-import { useEffect, useState } from 'react';
 
 import Checkbox from '@/components/common/Checkbox/Checkbox';
 import SelectionCounter from '@/components/common/SelectionCounter/SelectionCounter';
 import { totalOptionCount } from '@/components/NewChecklist/OptionModal/OptionModal';
+import useAllSelect from '@/hooks/useAllSelect';
 import { flexCenter, title4 } from '@/styles/common';
 import theme from '@/styles/theme';
 
@@ -12,29 +12,24 @@ interface Props {
   setSelectedOptions: React.Dispatch<number[]>;
 }
 const OptionModalInfoBox = ({ selectedOptions, setSelectedOptions }: Props) => {
-  const [isAllSelected, setIsAllSelected] = useState(selectedOptions.length === totalOptionCount);
-
-  const onClickSelectAllOptions = () => {
-    if (isAllSelected) {
-      return setSelectedOptions([]);
-    }
-    const allOptions = new Array(totalOptionCount).fill(0).map((e, i) => i + 1);
-    return setSelectedOptions(allOptions);
-  };
-
-  useEffect(() => {
-    setIsAllSelected(selectedOptions.length === totalOptionCount);
-  }, [selectedOptions]);
+  const allOptions = new Array(totalOptionCount).fill(0).map((e, i) => i + 1);
+  const { onClickSelectAllOptions, isAllSelected, setIsAllSelected } = useAllSelect({
+    allOptions,
+    setSelectedOptions,
+    selectedOptions,
+  });
 
   return (
     <S.ButtonContainer>
       <S.TotalSelectBox>
         <Checkbox
           isChecked={isAllSelected}
+          setIsChecked={setIsAllSelected}
           onClick={onClickSelectAllOptions}
           color={theme.palette.yellow500}
           hoverBorderColor={theme.palette.yellow600}
         />
+
         <span>전체선택</span>
       </S.TotalSelectBox>
       <SelectionCounter currentCount={selectedOptions.length} totalCount={totalOptionCount} />
