+++ frontend/src/hooks/useAllSelect.ts
@@ -1,9 +1,26 @@
-// import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 
-// const useAllSelect = )selectedOptions => {
-//   const [isAllSelected, setIsAllSelected] = useState(selectedOptions.length === totalOptionCount);
+interface Props {
+  allOptions: (number | string)[];
+  setSelectedOptions: React.Dispatch<(number | string)[]>;
+  selectedOptions: (number | string)[];
+}
+const useAllSelect = ({ allOptions, setSelectedOptions, selectedOptions }: Props) => {
+  const [isAllSelected, setIsAllSelected] = useState<boolean>(false);
 
-//   useEffect(() => {}, [selectedOptions]);
-// };
+  const onClickSelectAllOptions = () => {
+    if (isAllSelected) {
+      return setSelectedOptions([]);
+    }
 
-// export default useAllSelect;
+    return setSelectedOptions(allOptions);
+  };
+
+  useEffect(() => {
+    setIsAllSelected(selectedOptions.length === allOptions.length);
+  }, [selectedOptions]);
+
+  return { onClickSelectAllOptions, isAllSelected, setIsAllSelected };
+};
+
+export default useAllSelect;
