--- frontend/src/pages/ModalPage.tsx
@@ -1,5 +1,0 @@
-import OptionModal from '@/components/NewChecklist/OptionModal/OptionModal';
-
-export const ModalPage = () => {
-  return <OptionModal />;
-};
