+++ frontend/src/styles/global.ts
@@ -10,7 +10,6 @@ export const baseStyle = css`
     max-width: 600px;
     height: 100vh;
     margin: 0px auto;
-    outline: 1px solid ${theme.palette.grey200};
     outline-offset: -1px;
     box-shadow: 0 0 20px ${theme.palette.grey100};
   }
